ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
pandas==2.1.4
scikit-learn==1.4.0
transformers==4.36.2
openai==1.10.0
pymongo==4.6.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
prophet==1.1.5
matplotlib==3.8.2
xgboost==2.0.3
torch==2.1.2
numpy==1.26.3
pytest==7.4.4
aiohttp==3.9.1
cachetools==5.3.2
python-dateutil==2.8.2 